quiet a warning by coercing to unsigned since we know it is not lt 0 here

+++ src/Stack_string_base.cpp
@@ -106,7 +106,7 @@ int Stack_string_base::sprintf(const char *format, ...)
 		return ret;
 	}
 
-	if(ret <= char_len)
+	if(size_t(ret) <= char_len)
 	{
 		m_len += ret;
 	}
